Move framework code into gundi_action_runner; shim app/* for fork compatibility
