+++ app/actions/core.py
@@ -1,95 +1,11 @@
+"""Deprecated compatibility shim — this module moved to gundi_action_runner.actions.core."""
 import importlib
-import inspect
-from typing import Optional
+import sys
+import warnings
 
-from pydantic import BaseModel, Field
-from app.services.utils import UISchemaModelMixin
-
-
-class ActionConfiguration(UISchemaModelMixin, BaseModel):
-    pass
-
-
-class InternalActionConfiguration(BaseModel):
-    pass
-
-
-class PullActionConfiguration(ActionConfiguration):
-    # Pull actions are scheduled at the integration-type level, so the portal
-    # fires them for every integration of this type. This toggle lets an
-    # operator pause scheduled execution for a given integration without
-    # deleting its configuration — useful when the integration is used only as
-    # a destination and the pull is not intended to run. The action_runner also
-    # treats a missing or invalid pull config as a clean no-op rather than an
-    # error, so destination-only integrations stay quiet by default.
-    run_on_schedule: bool = Field(
-        True,
-        title="Run On Schedule",
-        description=(
-            "When enabled, this action runs automatically on its configured schedule. "
-            "Turn it off to pause scheduled execution for this integration without deleting "
-            "the configuration."
-        ),
-    )
-
-
-class ExecutableActionMixin:
-    pass
-
-
-def action_title(title: str):
-    """Set the display name used when registering the action in Gundi,
-    instead of the default derived from the handler function name."""
-    def decorator(func):
-        setattr(func, "action_title", title)
-        return func
-    return decorator
-
-
-class PushActionConfiguration(ActionConfiguration):
-    pass
-
-
-class AuthActionConfiguration(ActionConfiguration):
-    pass
-
-
-class GenericActionConfiguration(ActionConfiguration):
-    pass
-
-
-def discover_actions(module_name, prefix):
-    action_handlers = {}
-    # Import the module using importlib
-    module = importlib.import_module(module_name)
-    all_members = inspect.getmembers(module)
-
-    # Iterate through the members and filter functions by prefix
-    for name, func in all_members:
-        if name.startswith(prefix) and inspect.isfunction(func):
-            signature = inspect.signature(func)
-            key = name[len(prefix):]  # Remove prefix
-            if (config_annotation := signature.parameters.get("action_config").annotation) != inspect._empty:
-                config_model = config_annotation
-            else:
-                config_model = GenericActionConfiguration
-            if issubclass(config_model, PushActionConfiguration):  # Push actions
-                if data_param := signature.parameters.get("data"):
-                    if (data_annotation := data_param.annotation) != inspect._empty:
-                        data_model = data_annotation
-                    else:
-                        raise ValueError(f"The 'data' parameter in action '{key}' must be annotated with a data model.")
-                else:
-                    raise ValueError(f"Push action '{key}' must accept a 'data' parameter.")
-                if not signature.parameters.get("metadata"):
-                    raise ValueError(f"Push action '{key}' must accept a 'metadata' parameter.")
-            else:
-                data_model = None
-            action_handlers[key] = (func, config_model, data_model)
-
-    return action_handlers
-
-
-
-def get_actions():
-    return list(discover_actions(module_name="app.actions.handlers", prefix="action_").keys())
+warnings.warn(
+    "'app.actions.core' is deprecated; import 'gundi_action_runner.actions.core' instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+sys.modules[__name__] = importlib.import_module("gundi_action_runner.actions.core")
