+++ app/services/core.py
@@ -1,8 +1,11 @@
-from enum import Enum
+"""Deprecated compatibility shim — this module moved to gundi_action_runner.services.core."""
+import importlib
+import sys
+import warnings
 
-
-class ActionTypeEnum(str, Enum):
-    AUTHENTICATION = "auth"
-    PULL_DATA = "pull"
-    PUSH_DATA = "push"
-    GENERIC = "generic"
+warnings.warn(
+    "'app.services.core' is deprecated; import 'gundi_action_runner.services.core' instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+sys.modules[__name__] = importlib.import_module("gundi_action_runner.services.core")
