+++ src/gundi_action_runner/services/core.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class ActionTypeEnum(str, Enum):
+    AUTHENTICATION = "auth"
+    PULL_DATA = "pull"
+    PUSH_DATA = "push"
+    GENERIC = "generic"
