+++ src/gundi_action_runner/services/errors.py
@@ -1,0 +1,17 @@
+
+
+class ActionNotFound(Exception):
+    pass
+
+
+class ConfigurationNotFound(Exception):
+    pass
+
+
+class ConfigurationValidationError(Exception):
+    pass
+
+
+class ActionExecutionError(Exception):
+    pass
+
