+++ tests/test_action_runner.py
@@ -8,10 +8,10 @@ from gundi_core.commands import RunIntegrationAction
 from gundi_core.events import IntegrationActionFailed, IntegrationActionCustomLog, LogLevel
 from gundi_core.events.transformers import ObservationTransformedER
 
-from app import settings
-from app.conftest import MockSubActionConfiguration, MockPushActionConfiguration, async_return
+from gundi_action_runner import settings
+from gundi_action_runner.testing.fixtures import MockSubActionConfiguration, MockPushActionConfiguration, async_return
 from app.main import app
-from app.services.action_scheduler import trigger_action
+from gundi_action_runner.services.action_scheduler import trigger_action
 
 api_client = TestClient(app)
 
@@ -37,11 +37,11 @@ async def test_execute_pull_action_from_pubsub(
         mocker, mock_gundi_client_v2, mock_publish_event, mock_action_handlers, mock_config_manager,
         pubsub_message_request_headers, run_pull_action_pubsub_payload
 ):
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
 
     response = api_client.post(
         "/",
@@ -66,12 +66,12 @@ async def test_execute_push_action_from_pubsub(
         mocker, mock_gundi_client_v2, mock_publish_event, mock_action_handlers, mock_config_manager,
         pubsub_message_request_headers, run_push_action_pubsub_payload, mock_push_observations_handler
 ):
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.actions.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.actions.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
 
     response = api_client.post(
         "/push-data",
@@ -107,11 +107,11 @@ async def test_execute_action_from_api(
         mocker, mock_gundi_client_v2, integration_v2, mock_config_manager,
         mock_publish_event, mock_action_handlers,
 ):
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     integration_id = str(integration_v2.id)
     action_id = "pull_observations"
 
@@ -136,11 +136,11 @@ async def test_execute_action_from_api_with_config_overrides(
         mocker, mock_gundi_client_v2, integration_v2, mock_config_manager,
         mock_publish_event, mock_action_handlers,
 ):
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
 
     config_overrides = {"lookback_days": 3}
     response = api_client.post(
@@ -167,11 +167,11 @@ async def test_execute_action_from_pubsub_with_config_overrides(
         mocker, mock_gundi_client_v2, mock_publish_event, mock_action_handlers, mock_config_manager,
         pubsub_message_request_headers, run_pull_action_pubsub_payload_with_config_overrides
 ):
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
 
     response = api_client.post(
         "/",
@@ -199,11 +199,11 @@ async def test_manual_pull_action_with_invalid_config_still_errors(
 ):
     # A direct /execute call is a manual run → strict: invalid config 422s so
     # the operator sees the misconfiguration immediately.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
 
     response = api_client.post(
         "/v1/actions/execute/",
@@ -227,11 +227,11 @@ async def test_triggered_by_marker_is_case_insensitive(
     # A mixed-case "MANUAL" marker must be honored as a manual run (strict), not
     # silently fall through to the automated default. With an invalid config that
     # means it errors (IntegrationActionFailed) rather than skipping quietly.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     bad_config = mocker.MagicMock()
     bad_config.data = {"lookback_days": "two"}  # should be an integer
     mock_config_manager.get_action_configuration.return_value = async_return(bad_config)
@@ -262,12 +262,12 @@ async def test_scheduled_pull_action_with_invalid_config_is_skipped(
     # A scheduled (PubSub, no triggered_by → automated) pull whose stored config
     # is invalid skips cleanly: no handler call, NO IntegrationActionFailed, and
     # — when the throttle window is open — one WARNING activity log with detail.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.action_runner.state_manager", mock_state_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.action_runner.state_manager", mock_state_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     mock_state_manager.set_if_absent.return_value = async_return(True)  # window open
     bad_config = mocker.MagicMock()
     bad_config.data = {"lookback_days": "two"}  # should be an integer
@@ -296,12 +296,12 @@ async def test_scheduled_pull_action_invalid_config_warning_is_throttled(
     # When the throttle window is closed (set_if_absent → False), the skip is
     # still logged locally but NO portal WARNING is published — so a
     # persistently misconfigured source doesn't emit a warning every tick.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.action_runner.state_manager", mock_state_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.action_runner.state_manager", mock_state_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     mock_state_manager.set_if_absent.return_value = async_return(False)  # window closed
     bad_config = mocker.MagicMock()
     bad_config.data = {"lookback_days": "two"}
@@ -327,12 +327,12 @@ async def test_scheduled_pull_action_invalid_config_skip_survives_throttle_failu
     # If the throttle store (Redis) is unavailable, the skip must not crash the
     # request (which would 500 / trigger PubSub redelivery). It degrades open:
     # the WARNING is still published this time, and nothing is raised.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.action_runner.state_manager", mock_state_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.action_runner.state_manager", mock_state_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     mock_state_manager.set_if_absent.side_effect = Exception("redis unavailable")
     bad_config = mocker.MagicMock()
     bad_config.data = {"lookback_days": "two"}
@@ -360,11 +360,11 @@ async def test_scheduled_pull_action_with_missing_config_is_skipped(
     # Destination-only integrations have pull actions scheduled type-wide but no
     # pull config at all — an expected, quiet no-op: local log only, NO portal
     # activity-feed event at all.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     mock_config_manager.get_action_configuration.return_value = async_return(None)
 
     response = api_client.post(
@@ -385,11 +385,11 @@ async def test_scheduled_pull_action_skipped_when_run_on_schedule_disabled(
 ):
     # A valid config with run_on_schedule off pauses scheduled execution — also
     # a quiet, local-log-only skip with no portal activity-feed event.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     paused_config = mocker.MagicMock()
     paused_config.data = {"lookback_days": 10, "run_on_schedule": False}
     mock_config_manager.get_action_configuration.return_value = async_return(paused_config)
@@ -411,11 +411,11 @@ async def test_manual_pull_action_runs_even_when_run_on_schedule_disabled(
         mock_publish_event, mock_action_handlers,
 ):
     # The pause toggle only gates scheduled runs — a manual /execute still runs.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     paused_config = mocker.MagicMock()
     paused_config.data = {"lookback_days": 10, "run_on_schedule": False}
     mock_config_manager.get_action_configuration.return_value = async_return(paused_config)
@@ -440,11 +440,11 @@ async def test_non_pull_action_still_errors_on_invalid_config(
 ):
     # The skip-on-invalid behavior is scoped to pull actions only — a non-pull
     # (here InternalActionConfiguration) action with a bad config still 422s.
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
     bad_config = mocker.MagicMock()
     bad_config.data = {"start_datetime": "not-a-datetime", "end_datetime": "also-bad"}
     mock_config_manager.get_action_configuration.return_value = async_return(bad_config)
@@ -469,12 +469,12 @@ async def test_trigger_subaction(
 ):
     settings.TRIGGER_ACTIONS_ALWAYS_SYNC = False
     settings.INTEGRATION_COMMANDS_TOPIC = "integration-actions-topic"
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_scheduler.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_scheduler.publish_event", mock_publish_event)
     integration_id = str(integration_v2.id)
     action_id = "pull_observations_by_date"
     config = MockSubActionConfiguration(
@@ -508,12 +508,12 @@ async def test_trigger_subaction_sync(
         mock_publish_event, mock_action_handlers,
 ):
     settings.TRIGGER_ACTIONS_ALWAYS_SYNC = True
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_scheduler.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_scheduler.publish_event", mock_publish_event)
     integration_id = str(integration_v2.id)
     action_id = "pull_observations_by_date"
     config = MockSubActionConfiguration(
@@ -543,11 +543,11 @@ async def test_execute_action_with_handler_error(
         mocker, mock_gundi_client_v2, integration_v2, mock_config_manager,
         mock_publish_event, mock_action_handlers_with_request_errors
 ):
-    mocker.patch("app.services.action_runner.action_handlers", mock_action_handlers_with_request_errors)
-    mocker.patch("app.services.action_runner._portal", mock_gundi_client_v2)
-    mocker.patch("app.services.action_runner.config_manager", mock_config_manager)
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    mocker.patch("app.services.action_runner.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.action_handlers", mock_action_handlers_with_request_errors)
+    mocker.patch("gundi_action_runner.services.action_runner._portal", mock_gundi_client_v2)
+    mocker.patch("gundi_action_runner.services.action_runner.config_manager", mock_config_manager)
+    mocker.patch("gundi_action_runner.services.activity_logger.publish_event", mock_publish_event)
+    mocker.patch("gundi_action_runner.services.action_runner.publish_event", mock_publish_event)
 
     response = api_client.post(
         "/v1/actions/execute/",
